split comparison into it's own component
